+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Removes .NET Framework 4.5.2 support. The minimum .NET Framework
   version supported is .NET 4.6.1. ([#2138](https://github.com/open-telemetry/opentelemetry-dotnet/issues/2138))
 
+* `HttpClient` instances created before `AddHttpClientInstrumentation` is called
+  on .NET Framework will now also be instrumented
+  ([#2364](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2364))
+
 ## 1.0.0-rc7
 
 Released 2021-Jul-12